AssetsBtnsList as separate component
